fix(sentinel): test-event handler fills fingerprint + payload
Root cause: BigInteger PK breaks SQLite autoincrement (BIGINT != INTEGER PRIMARY KEY).
- sentinel.py models: BigInteger → Integer on all PKs and FK cols
- migration.py: _recreate_sentinel_tables() drops old tables and recreates
  with INTEGER PRIMARY KEY AUTOINCREMENT on first boot after deploy
- migration gated by schema_migrations record so it runs exactly once

+++ backend/app/db/models/sentinel.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from datetime import datetime, timezone
 from typing import Optional
 
-from sqlalchemy import BigInteger, Boolean, DateTime, Integer, Numeric, String, Text
+from sqlalchemy import Boolean, DateTime, Integer, Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -17,7 +17,7 @@ def _now() -> datetime:
 class AgentEvent(Base):
     __tablename__ = "agent_events"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     agent_id: Mapped[str] = mapped_column(String, nullable=False, index=True)
     severity: Mapped[str] = mapped_column(String, nullable=False)
     category: Mapped[str] = mapped_column(String, nullable=False, index=True)
@@ -31,8 +31,8 @@ class AgentEvent(Base):
 class AgentFix(Base):
     __tablename__ = "agent_fixes"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    event_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    event_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
     fixer_agent: Mapped[str] = mapped_column(String, nullable=False)
     pr_number: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     pr_url: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
@@ -47,7 +47,7 @@ class AgentFix(Base):
 class AgentCircuitBreaker(Base):
     __tablename__ = "agent_circuit_breakers"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     breaker_type: Mapped[str] = mapped_column(String, nullable=False, index=True)
     key: Mapped[str] = mapped_column(String, nullable=False, index=True)
     tripped_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, default=_now)
@@ -58,8 +58,8 @@ class AgentCircuitBreaker(Base):
 class AgentEscalation(Base):
     __tablename__ = "agent_escalations"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    event_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    event_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
     paste_ready: Mapped[str] = mapped_column(Text, nullable=False)
     sent_to_discord: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
     sent_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), nullable=True)
